ci: add nix env and gha workflows
First pass at a dev env, which pulls in the usual Rust deps, as well as
go, which is required to build `penumbra-indexer`. The devshell pulls in
the necessary deps, resulting in `cargo build` running the
proxy build for go via `build.rs` just fine.

Not yet supported is `nix build`: will need to figure out a two-pass
approach, because network access during the buildPhase is not permitted,
so the go build via rust fails.

CI cache timing info:

  * First run of cargo-check: 3m13s
  * First run of cargo-fmt: 1m57s