build: support 'nix build' for cometbft archive
The existing build.rs script assumes that go has been set up
sufficiently, and calls out to it to build the static library for
cometbft for linking against. That's rad, but naively calling out to
`go build` from inside `nix build` fails, because the rust
overlay doesn't know anything about golang, and forbids arbitrary
network calls.

Instead, we'll have to be explicit about preparing the `libcomebft.a` static lib
as a discrete output, and consuming it as an input in the rust build.
We can do so by opting in to using a prebuilt lib via env var in the
`build.rs` script.

This change improves the nix config, making it sufficiently defined to
perform a `nix build` successfully. Careful attention was given to
preserving the existing functionality so that `cargo build` will still
do the right thing, as long as golang is available.