feat: custom compatibility type for BeginBlock
Not finished yet, but I think this is the pattern we're forced to use.
We cannot round-trip via serde because the tendermint crates use
protobuf for serialization, which means versioning info is preserved,
prohibiting naive round-tripping across versions. So instead I'm
manually unpacking every field, reducing it to primitive types, and
re-encoding.