Fix parsing bugs in begin block (#35)
* Bug: incorrect compatability conversion of timestamps

Turns out that nanoseconds are indeed the whole horse. We need to
instead truncate them, to make the conversion not fail. This is kind
of unnecessary because in practice our timestamps do not have this
granularity, but, good to avoid data loss at this layer, which doesn't
know this.

* Bug: correctly handle Hash conversion.

Turns out we weren't handling the case of Hash::None correctly!

* Add test harness for begin block parsing

Can be run with:
```
cargo test --features expensive-tests --test file '' -- <ARCHIVE> 
```

* fix: feature-gate beginblock test

We don't want to run the beginblock test on default test invocations,
only if the `expensive-tests` feature is enabled.

* tests: use env var for fixture filepath

Rather than using cli arg-parsing, which breaks on alternative `cargo
test` invocations, let's require the filepath to be set via env var.
Clunky, but durable. Added a justfile target that reuses an existing
sqlite3 db, created by the integration test suite.

* fix: default logging at info in tests

Updates the `init_tracing` fn from `src/lib.rs`, and
adds a bit of informative logging to the new beginblock test.

* make plain `cargo test` pass

This will make the tests that need arguments be allowed to fail.

---------

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>