Allow regen to follow a remote node
This augments the regen command to take a `--follow` input,
which will allow it to run *past* the blocks in the archive, using
the remote node as a source of new blocks. These blocks will be
put inside the archive as well.

This command will first exhaust the archive before fetching remote blocks.

If there's no end block, this will poll the remote store to run forever.