Allow regen to follow a remote node (#43)
* Allow RemoteStore to stream new blocks

* Allow regen to follow a remote node

This augments the regen command to take a `--follow` input,
which will allow it to run *past* the blocks in the archive, using
the remote node as a source of new blocks. These blocks will be
put inside the archive as well.

This command will first exhaust the archive before fetching remote blocks.

If there's no end block, this will poll the remote store to run forever.

* Add better logging when following remote node

* docs: more info in regen follow description

---------

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>