Simplify and improve remote block streaming
One thing I noticed when syncing up from very far, is that you'd
first request the height bounds, download a lot of blocks, regenerating,
and then you'd move into the block-by-block flow. But then,
you'd actually be hundreds and hundreds of blocks behind, so this was
artificially slow, when there were in fact many blocks available
on the remote, but you were downloading them one at a time!

Now, this simplifies the loop to take a pattern of:
- (outer loop) polling to see if there are new blocks on the remote
- (inner loop) streaming all the new blocks

This unifies both the "catch up" and the "follow" phases.