feat: fancy progress bars
Used claude to add some visual progress bars for interactive sessions.
The primary utility here is that there's an ETA displayed, which is
super helpful both for large file downloads, and for remote block
fetching, which only pulls at a rate of ~40 blocks per second, which
falls short of 10x network speed.