feat: automatic "regen" subcommand
Updates the "regen" logic to infer the regeneration plan based on chain
id, and run the regeneration logic for each relevant protocol version.
This removes the need to wrap the regen invocation with a script to feed
in all relevant `--stop-height` flags.

In a nutshell, this change:

  * renames "regen" -> "regen-step"
  * adds a new "regen" subcmd that reads RegenerationPlan
  * shells out via current_exe to "regen-step" according to plan

The trick of shelling out via current_exe is employed to sidestep the
rough edge in the upstream Penumbra logic that calls sys-exit on upgrade
boundaries, which will cause the `penumbra-reindexer` process to
terminate, too, unless we shell out to it. This is a bit of a hack, but
in the interest of ergonomics, we can sweep the complexity into an
internal-only subcommand by "hiding" the "regen-step" logic via clap.

Includes minor updates to README to match.